Renamed entities to schemas
